Added style props for Calendar
For the case that you need to change style of Calendar component.
For example, if you want to change position or with of the component, you can simply set style like this 
`{ position: 'absolute', left: 0, top: 0, width: '290px' }`

This component is not displayed well in my project for some reason when I use with Input. So I need a little customization. Thx